% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscyclic.R
\name{is.cyclic}
\alias{is.cyclic}
\title{Checks if matrix is cyclic}
\usage{
is.cyclic(mat)
}
\arguments{
\item{mat}{adjacency matrix of graph (values in 1, 2, 3)}
}
\value{

}
\description{
Checks if matrix is cyclic
}
\examples{
mat <- matrix(c(1,0,0,1), byrow=TRUE)
is.cyclic(mat)
}
