% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvida.R
\name{lv.ida.lm}
\alias{lv.ida.lm}
\title{Run Latent Variable-IDA - Linear Models}
\usage{
lv.ida.lm(
  x.pos,
  y.pos,
  data,
  pag,
  method = "global",
  nMags = 500,
  localcap = NULL,
  possdsep = "small",
  verbose = FALSE,
  mags.local = FALSE,
  Z_i = NULL,
  opag = pag,
  bugwatch = FALSE
)
}
\arguments{
\item{x.pos}{name of column of x in d.mat}

\item{y.pos}{name of column of y in d.mat}

\item{data}{data matrix or data frame (with named columns)}

\item{pag}{PAG matrix found by FCI (with named rows and columns)**this is necessary for the time being because there isn't a way to get from the graphical object to the matrix**}

\item{method}{"local" - local; "global" - all MAGs. Deprecated.}

\item{verbose}{Verbose logging output}

\item{bugwatch}{}
}
\value{
causal values
}
\description{
Estimate the causal effect of x on y; no precomputed covariance matrix required (so can handle factors); all MAGs can be precomputed
}
