% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvida.R
\name{lv.ida}
\alias{lv.ida}
\title{Estimate Causal Effects with LV-IDA}
\usage{
lv.ida(
  x.pos,
  y.pos,
  pag,
  data = NULL,
  mcov = NULL,
  method = "global",
  nMags = 500,
  localcap = NULL,
  possdsep = "small",
  verbose = FALSE,
  mags.local = FALSE,
  Z_i = NULL,
  opag = pag,
  bugwatch = FALSE
)
}
\arguments{
\item{x.pos}{}

\item{y.pos}{}

\item{pag}{}

\item{data}{}

\item{mcov}{}

\item{method}{}

\item{nMags}{}

\item{localcap}{}

\item{possdsep}{}

\item{verbose}{}

\item{mags.local}{}

\item{Z_i}{}

\item{opag}{}

\item{bugwatch}{}
}
\value{

}
\description{
Estimates the effect p(Y | do(X)) in a PAG using the Latent-Variable IDA algorithm by Malinsky and Spirtes.
}
